refactor: move plugins to a dedicated folder
This is an internal change only, designed to facilitate the organization of plugins. There is no impact for maxGraph users.

In addition:
  - use "type" imports for plugin when possible
  - remove duplication when calling plugins
