+++ packages/core/src/editor/Editor.ts
@@ -47,14 +47,14 @@ import MaxLog from '../gui/MaxLog';
 import { isNode } from '../util/domUtils';
 import { getViewXml, getXml } from '../util/xmlUtils';
 import { load, post, submit } from '../util/MaxXmlRequest';
-import PopupMenuHandler from '../view/handler/PopupMenuHandler';
-import RubberBandHandler from '../view/handler/RubberBandHandler';
+import type PopupMenuHandler from '../view/plugins/PopupMenuHandler';
+import RubberBandHandler from '../view/plugins/RubberBandHandler';
 import InternalEvent from '../view/event/InternalEvent';
 import InternalMouseEvent from '../view/event/InternalMouseEvent';
 import { CellStateStyle, MouseListenerSet } from '../types';
-import ConnectionHandler from '../view/handler/ConnectionHandler';
+import type ConnectionHandler from '../view/plugins/ConnectionHandler';
 import { show } from '../util/printUtils';
-import PanningHandler from '../view/handler/PanningHandler';
+import type PanningHandler from '../view/plugins/PanningHandler';
 import { cloneCell } from '../util/cellArrayUtils';
 import { TranslationsConfig } from '../i18n/config';
 import type MaxPopupMenu from '../gui/MaxPopupMenu';
