+++ packages/core/src/view/mixins/ConnectionsMixin.ts
@@ -23,7 +23,7 @@ import EventObject from '../event/EventObject';
 import InternalEvent from '../event/InternalEvent';
 import Dictionary from '../../util/Dictionary';
 import type { Graph } from '../Graph';
-import type ConnectionHandler from '../handler/ConnectionHandler';
+import type ConnectionHandler from '../plugins/ConnectionHandler';
 
 type PartialGraph = Pick<Graph, 'getView' | 'getDataModel' | 'isPortsEnabled'>;
 type PartialConnections = Pick<
