+++ packages/core/src/view/plugins/PanningHandler.ts
