+++ packages/core/src/view/plugins/RubberBandHandler.ts
